# others/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FIAP Cloud Games API",
    "description": "API para gerenciamento de jogos, biblioteca de usuários e promoções",
    "contact": {
      "name": "FIAP Cloud Games Team",
      "email": "support@fiapcloudgames.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Library/user/{userId}": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Obtém a biblioteca completa de um usuário",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna a biblioteca do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Library"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Erro na solicitação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/{id}": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Obtém uma entrada específica da biblioteca pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da entrada da biblioteca",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna a entrada da biblioteca",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "404": {
            "description": "Entrada não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Erro na solicitação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Library/purchase": {
      "post": {
        "tags": [
          "Library"
        ],
        "summary": "Realiza a compra de um jogo para um usuário",
        "requestBody": {
          "description": "Dados da compra",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseGameRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseGameRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseGameRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Compra realizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflito - usuário já possui o jogo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Obtém um usuário pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do usuário",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Library": {
        "required": [
          "gameId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "gameId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "PurchaseGameRequest": {
        "required": [
          "gameId",
          "purchasePrice",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "description": "ID do usuário que está comprando o jogo",
            "format": "int32"
          },
          "gameId": {
            "type": "integer",
            "description": "ID do jogo a ser comprado",
            "format": "int32"
          },
          "purchasePrice": {
            "minimum": 0,
            "type": "number",
            "description": "Preço pago pelo jogo",
            "format": "double"
          },
          "isGift": {
            "type": "boolean",
            "description": "Indica se a compra é um presente"
          },
          "giftMessage": {
            "type": "string",
            "description": "Mensagem do presente (opcional)",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Modelo de requisição para compra de jogos"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Informe o token JWT no formato: Bearer {seu token}",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}